Compute selectedVehicle and selectedRoutes inside App
